Fix up polymorphism for core model types
Previously the manuscript, holding, and record classes weren't doing much.
They were added with an eye toward polymorphism, but the ExportRepresenter
was converting all item hashes unconditionally to a generic DsItem object
with methods applicable to all three.

This change gets ExportRepresenter working as intended so that Holding,
Manuscript, and Record objects are created by ExportRepresenter where
appropriate. DsItem is removed as no longer needed, and all unused methods
of the core model types are removed as well.

The spec helper export hash is also updated to accommodate these changes.

Addresses comment on #3